jiraairca/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
transformers
torch
jira
requests
atlassian-python-api